requirements.txt
aiogram==3.2.0 
sqlalchemy[asyncio]==2.0.23 
asyncpg==0.29.0 
alembic==1.13.0 
python-dotenv==1.0.0